Refactor reducers’s file names under Surveys screen (additional changes)
Replace singulars with plurals.
