+++ src/screens/Surveys/screens/SurveyList/data/reducers.js
@@ -2,7 +2,7 @@
 import { combineReducers } from 'redux';
 
 // Reducers
-import surveysReducer from './surveys/reducer';
+import surveysReducer from './surveys/reducers';
 
 // Combine reducers
 export default combineReducers({
