+++ src/screens/Surveys/screens/SurveyList/index.js
@@ -6,7 +6,7 @@ import * as commonActions from '../../actions';
 
 // Peer dependencies
 import * as dataActions from './data/surveys/actions';
-import { getSurveys } from './data/surveys/reducer';
+import { getSurveys } from './data/surveys/reducers';
 import * as uiActions from './actions';
 import { getUI, getView } from './reducers';
 import UI from './UI';
